requirements.txt
openai
speechrecognition
pyttsx3
pyaudio
tk
requests
python-dotenv   